Create style.css
This is the CSS file of the fashion project .It includes basic design elements such as font styling, text color, and alignment to enhance the readability and visual appeal of the content. The file also ensures proper spacing, organizes the layout, and adds borders around images to create a neat look.